condsolidated datastore key generation to DatastoreKeyUtil.java
Centralized key generation and also made documents children of their respective status entities within Datastore. This should allow for easier querying for all documents related to a single document ID.
